Assistant checkpoint: Fix ESLint errors and add missing Babel plugin
Assistant generated file changes:
- src/App.js: Fix ESLint errors and reorganize imports

---

User prompt:

Failed to compile.

[eslint]
src/App.js
  Line 10:1:  Import in body of module; reorder to top  import/first
  Line 11:1:  Import in body of module; reorder to top  import/first

Search for the keywords to learn more about each error.
WARNING in [eslint]
src/App.js
  Line 5:1:    Date prototype is read only, properties should not be added  no-extend-native
  Line 20:17:  'setLimit' is assigned a value but never used                no-unused-vars

ERROR in [eslint]
src/App.js
  Line 10:1:  Import in body of module; reorder to top  import/first
  Line 11:1:  Import in body of module; reorder to top  import/first

Search for the keywords to learn more about each error.

webpack compiled with 1 error and 1 warning
One of your dependencies, babel-preset-react-app, is importing the
"@babel/plugin-proposal-private-property-in-object" package without
declaring it in its dependencies. This is currently working because
"@babel/plugin-proposal-private-property-in-object" is already in your
node_modules folder for unrelated reasons, but it may break at any time.

babel-preset-react-app is part of the create-react-app project, which
is not maintianed anymore. It is thus unlikely that this bug will
ever be fixed. Add "@babel/plugin-proposal-private-property-in-object" to
your devDependencies to work around this error. This will make this message
go away.